@@ -3,10 +3,12 @@
     <section class="dark-gray-bg">
       <div class="container header">
         <div class="centered">
-          <!-- <h5 class="accent">{{category}}</h5> -->
-          <h2 class="white">{{ title }}</h2>
-          <div class="description gray">{{ author }}</div>
-          <div class="description white">{{ publish_on | formatDate }}</div>
+          <!-- <h5 class="accent">{{article.category}}</h5> -->
+          <h2 class="white">{{ article.title }}</h2>
+          <div class="description gray">{{ article.author }}</div>
+          <div class="description white">
+            {{ $filters.formatDate(article.publish_on) }}
+          </div>
         </div>
       </div>
     </section>
@@ -19,7 +21,7 @@
           color="accent"
           direction="left"
         />
-        <div v-html="body"></div>
+        <div v-html="article.body"></div>
         <v-more
           msg="Back to News"
           link="/news"
@@ -40,7 +42,7 @@
       :category="related.category"
       :title="related.title"
       :by="related.author"
-      :date="related.publish_on | formatDate"
+      :date="$filters.formatDate(related.publish_on)"
       :text="related.summary"
     />
   </div>
@@ -48,70 +50,92 @@
 
 <script>
 import moment from "moment";
+import { onBeforeRouteUpdate, useRouter, useRoute } from "vue-router";
+import { ref } from "vue";
+const api = require("../api");
 
 export default {
   name: "v-news",
-  data() {
-    return {
-      category: "",
-      title: "",
-      author: "",
-      publish_on: "",
-      body: "",
-      related: {
-        id: "",
-        category: "",
-        title: "",
-        author: "",
-        publish_on: "",
-        intro: ""
+  setup() {
+    const router = useRouter();
+    const route = useRoute();
+    const article = ref([]);
+    const related = ref([]);
+    const baseURL = ref("");
+
+    load(event, route.params.id);
+    async function load(event, id) {
+      try {
+        baseURL.value = api.getUri();
+        const newsItem = await api.get("/items/news/" + id, {
+          params: {
+            "fields[]": "*,related_article.*",
+            "filter[status][_eq]": "published",
+            "filter[publish_on][_lte]": moment().format("YYYY-MM-DD HH:mm:ss"),
+          },
+        });
+        article.value = newsItem.data.data;
+        related.value = newsItem.data.data.related_article;
+        // console.log({ newsItem: newsItem });
+        // console.log({ article: article });
+        // console.log({ related: related });
+      } catch (err) {
+        // eslint-disable-next-line
+        console.log('Error fetching "News Article"', err);
+        router.push("/not-found");
       }
+    }
+    //    });
+    onBeforeRouteUpdate(async (to, from, next) => {
+      load(event, to.params.id);
+      next();
+    });
+    return {
+      article,
+      related,
+      load,
+      baseURL,
     };
   },
-  beforeRouteUpdate(to, from, next) {
-    this.load(event, to.params.id);
-    next();
-  },
-  created: function () {
-    this.load(event, this.$route.params.id);
-  },
-  methods: {
-    load: function (event, id) {
-      this.$api
-        .getItem("news", id, {
-          fields: "*,related_article.*",
-          "filter[status][eq]": "published",
-          "filter[publish_on][leq]": moment().format("YYYY-MM-DD HH:mm:ss")
-        })
-        .then(
-          function(res) {
-            this.category = res.data.category;
-            this.title = res.data.title;
-            this.author = res.data.author;
-            this.publish_on = res.data.publish_on;
-            this.body = res.data.body;
+  // methods: {
+  //   load: function (event, id) {
+  //     this.$api
+  //       .get("/items/news/" + id, {
+  //         params: {
+  //           "fields[]": "*,related_article.*",
+  //           "filter[status][_eq]": "published",
+  //           "filter[publish_on][_lte]": moment().format("YYYY-MM-DD HH:mm:ss"),
+  //         },
+  //       })
+  //       .then(
+  //         function (res) {
+  //           this.category = res.data.category;
+  //           this.title = res.data.title;
+  //           this.author = res.data.author;
+  //           this.publish_on = res.data.publish_on;
+  //           this.body = res.data.body;
 
-            this.related.id = res.data.related_article.id;
-            this.related.category = res.data.related_article.category;
-            this.related.title = res.data.related_article.title;
-            this.related.author = res.data.related_article.author;
-            this.related.publish_on = res.data.related_article.publish_on;
-            this.related.summary = res.data.related_article.summary;
-            // eslint-disable-next-line
-          }.bind(this)
-        )
-        .catch(
-          function() {
-            this.$router.push("/not-found");
-          }.bind(this)
-        );
-    }
-  }
+  //           this.related.id = res.data.related_article.id;
+  //           this.related.category = res.data.related_article.category;
+  //           this.related.title = res.data.related_article.title;
+  //           this.related.author = res.data.related_article.author;
+  //           this.related.publish_on = res.data.related_article.publish_on;
+  //           this.related.summary = res.data.related_article.summary;
+  //           // eslint-disable-next-line
+  //         }.bind(this)
+  //       )
+  //       .catch(
+  //         function () {
+  //           this.$router.push("/not-found");
+  //         }.bind(this)
+  //       );
+  //   },
+  // },
 };
 </script>
 
 <style lang="scss">
-@import "../assets/_variables.scss";
+@import "~@/assets/_variables.scss";
 .news-detail {
   .header {
     margin-top: 120px;
